use colors in custom classif
